backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
sqlalchemy==2.0.45
python-jose[cryptography]==3.3.0
passlib[argon2]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
better-exceptions==0.3.2
httpx==0.25.2
starlette==0.27.0
argon2-cffi==25.1.0